move component tests into components folder
